rename test files to be test_* instead of *_test to align with pytest conventions
